requirements.txt
Django
djangorestframework
psycopg2-binary
celery
redis
qdrant-client
python-dotenv
dj-database-url
gunicorn
Pillow
pytesseract
transformers
sentence-transformers
djangorestframework-simplejwt
PyPDF2
python-magic
pydub
SpeechRecognition
google-genai
google-generativeai
groq

